Fix score

+++ client/src/client/app/course/class/class-list/class-list-dialog.component.ts
@@ -82,11 +82,17 @@ export class ClassListDialog extends BaseComponent implements OnInit {
     }
 
     delete() {
-        if (this.selectedClass)
-        this.confirm('Are you sure to delete ?', () => {
-            this.selectedClass.deleteClass(this).subscribe(() => {
-                this.loadClasses();
+        if (this.selectedClass) {
+            CourseMember.listByClass(this, this.selectedClass.id).subscribe((members)=> {
+                if (members.length)
+                    this.error('You cannot delete class with member inside');
+                else
+                    this.confirm('Are you sure to delete ?', () => {
+                        this.selectedClass.deleteClass(this).subscribe(() => {
+                            this.loadClasses();
+                        })
+                    });
             })
-        });
+        }
     }
 }
